refactor(02-01): extract database.rs into database/ directory module
- Split monolithic 2829-line database.rs into 10 focused sub-modules
- mod.rs: SqliteDatabase struct, new(), get_connection(), MIGRATED_DBS, re-exports
- schema.rs: SCHEMA const, SessionUpdate struct
- session.rs: Session CRUD (update_session, update_session_tx, archive_session, etc.)
- daily.rs: Daily stats queries (get_today_total, get_today_token_total, etc.)
- monthly.rs: Monthly stats queries (get_month_total, session_active_in_month, etc.)
- context.rs: Context learning CRUD (get/insert/update/delete learned contexts)
- maintenance.rs: MaintenanceResult, perform_maintenance, should_vacuum, update_last_vacuum
- analytics.rs: SessionWithModel, cross-table queries (get_all_sessions, is_healthy, etc.)
- sync.rs: Turso-sync upserts and counts (feature-gated)
- tests.rs: All 17 existing unit tests relocated
- Public API preserved via pub use re-exports in mod.rs
- All 245 lib tests pass, all integration tests pass
- No changes to any file outside src/database/
